Implement IBMCloud client in terms of the new HTTPClient
