+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2025-01-17T10:05:11Z",
+  "generated_at": "2025-01-17T10:23:47Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -81,7 +81,7 @@
       {
         "hashed_secret": "fb34629c9af1ed4045b5d6f287426276b2be3a1e",
         "is_verified": false,
-        "line_number": 53,
+        "line_number": 43,
         "type": "Secret Keyword",
         "verified_result": null
       }
