+++ src/tests/test_using_example_data.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::client::{AppConfigurationClient, AppConfigurationClientIBMCloud};
+use crate::client::{AppConfigurationClient, AppConfigurationClientHttp};
 use crate::tests::TrivialEntity;
 use rstest::*;
 
@@ -20,7 +20,7 @@ use super::client_enterprise;
 use crate::{Feature, Property, Value};
 
 #[rstest]
-fn test_get_a_specific_feature(client_enterprise: AppConfigurationClientIBMCloud) {
+fn test_get_a_specific_feature(client_enterprise: AppConfigurationClientHttp) {
     let specific_feature = client_enterprise.get_feature_proxy("f1").unwrap();
 
     let name = specific_feature.get_name().unwrap();
@@ -33,7 +33,7 @@ fn test_get_a_specific_feature(client_enterprise: AppConfigurationClientIBMCloud
 }
 
 #[rstest]
-fn test_get_a_specific_property(client_enterprise: AppConfigurationClientIBMCloud) {
+fn test_get_a_specific_property(client_enterprise: AppConfigurationClientHttp) {
     let property = client_enterprise.get_property_proxy("p1").unwrap();
 
     let name = property.get_name().unwrap();
